Fix cascade issue with CSS Stylesheet
In order to make my code produce the desired
outcome, I made many changes to the css. I left the HTML file
untouched, and the css styles untouched as well, I just changed
either the selector type or style. I changed the specificity
of the small paragraph by making it a chain selector, I changed
the "confirm-button" to an ID instead of a class, and I also added
specificity to the "div.text" part by adding a type selector in front
of it.